flasks/sockjstornado_/requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.19
Werkzeug==0.9.4
itsdangerous==0.23

tornado==3.2
sockjs-tornado==1.0.0
